requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
Jinja2
python-dotenv
gTTS
assemblyai
langchain
langchain-google-genai
python-multipart
aiofiles
psycopg2-binary==2.9.9